move task part
