+++ src/task/spec.rs
@@ -7,7 +7,7 @@
 use std::fmt;
 use std::time::Duration;
 
-use crate::task::task::TaskRef;
+use crate::task::base::TaskRef;
 use crate::{config::Config, policy::RestartPolicy, strategy::BackoffStrategy};
 
 /// # Specification for running a task under supervision.
